% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_anno}
\alias{add_anno}
\title{A ggproto for annotating the symmetry of the starting points}
\usage{
add_anno(dt, anno_color = "black", anno_lty = "dashed", anno_alpha = 0.1, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{anno_color}{the color of the annotation}

\item{anno_lty}{the linetype of the annotation}

\item{anno_alpha}{an alpha value for the transparency of the annotation}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_center}()},
\code{\link{add_finish}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
