% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_end}
\alias{add_end}
\title{A ggproto for drawing start points}
\usage{
add_end(dt, end_size = 5, end_alpha = 1, end_color = NULL, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{end_size}{the point size}

\item{end_alpha}{an alpha value for the transparency of the point}

\item{end_color}{the color of the points}

\item{...}{other aesthetics inherent from \code{explore_space_pca()}}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_dir_search}()},
\code{\link{add_finish}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
