% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_finish}
\alias{add_finish}
\title{A ggproto for drawing finish points}
\usage{
add_finish(dt, finish_size = 0.5, finish_alpha = 1, finish_color = NULL, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{finish_size}{the point size}

\item{finish_alpha}{an alpha value for the transparency of the point}

\item{finish_color}{the color of the points}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
