% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_interp}
\alias{add_interp}
\title{A ggproto for drawing interpolation path}
\usage{
add_interp(
  dt,
  interp_size = 1.5,
  interp_alpha = NULL,
  interp_color = NULL,
  interp_group = NULL,
  ...
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{interp_size}{the size of the path}

\item{interp_alpha}{an alpha value for the transparency of the path}

\item{interp_color}{the color of the path}

\item{interp_group}{a group variable for path connection}

\item{...}{other argument passed to \code{draw_path()}}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
