% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_interrupt}
\alias{add_interrupt}
\title{A ggproto for annotating the interrupted path}
\usage{
add_interrupt(
  dt,
  interrupt_size = 0.5,
  interrupt_alpha = NULL,
  interrupt_color = NULL,
  interrupt_group = NULL,
  interrupt_linetype = "dashed",
  ...
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{interrupt_size}{the size of the path}

\item{interrupt_alpha}{an alpha value for the transparency of the path}

\item{interrupt_color}{the color of the path}

\item{interrupt_group}{a group variable for path connection}

\item{interrupt_linetype}{the linetype for annotating the interrupted path}

\item{...}{other aesthetics inherent from \code{explore_space_pca()}}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_dir_search}()},
\code{\link{add_end}()},
\code{\link{add_interp_last}()},
\code{\link{add_interp}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
