% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_search}
\alias{add_search}
\title{A ggproto for drawing search points}
\usage{
add_search(dt, search_size = 0.5, search_alpha = 1, search_color = NULL, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{search_size}{the point size}

\item{search_alpha}{an alpha value for the transparency of the point}

\item{search_color}{the color of the points}

\item{...}{other aesthetics inherent from \code{explore_space_pca()}}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_finish}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
