% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_theo}
\alias{add_theo}
\title{A ggproto for drawing the theoretical basis, if applicable}
\usage{
add_theo(dt, theo_label = "*", theo_size = 25)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{theo_label}{the symbol used for labelling the theoretical basis}

\item{theo_size}{the size of the label}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()}
}
\concept{draw functions}
