% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_random}
\alias{bind_random}
\title{Bind random bases in the projection bases space}
\usage{
bind_random(dt, n = 500)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{n}{Number of random points to generate in each dimension}
}
\description{
Given the orthonormality constraint, the projection bases live in a high dimensional hollow sphere.
Generating random points on the sphere is useful to preceive the data object in the high dimensional space.
}
\examples{
bind_random(holes_1d_better) \%>\% tail(5)
}
\seealso{
Other bind: 
\code{\link{bind_random_matrix}()},
\code{\link{bind_theoretical}()}
}
\concept{bind}
