% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_random_matrix}
\alias{bind_random_matrix}
\title{Bind random bases in the projection bases space as a matrix}
\usage{
bind_random_matrix(basis, n = 500)
}
\arguments{
\item{basis}{A basis matrix returned by \code{get_basis_matrix()}}

\item{n}{Number of random points to generate in each dimension}
}
\value{
matrix
}
\description{
Bind random bases in the projection bases space as a matrix
}
\examples{
data <- get_basis_matrix(holes_1d_geo)
bind_random_matrix(data) \%>\% tail(5)
}
\seealso{
Other bind: 
\code{\link{bind_random}()},
\code{\link{bind_theoretical}()}
}
\concept{bind}
