% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_theoretical}
\alias{bind_theoretical}
\alias{bind_random}
\title{bind external data to a given data object}
\usage{
bind_theoretical(dt, matrix, index)

bind_random(dt, ...)
}
\arguments{
\item{dt}{the data object being binded to}

\item{matrix}{the theoretical basis to bind}

\item{index}{the index function used to calculate index value}

\item{...}{Additional parameters pass to geozoo::sphere.hollow()}
}
\description{
These functions allows you to bind the theoretical best basis and randomly generated basis on a hollow sphere to the data object
}
\details{
In a simulation setting, the theoretical best basis is known for a given problem.

Given the orthonormality constraint, the projection bases live in a high dimensional hollow sphere.
Generating random points on the sphere is useful to preceive the data object in the high dimensional space.
}
\examples{
best <- matrix(c(0, 1, 0, 0, 0), nrow = 5)
holes_1d_better \%>\% bind_theoretical(best, tourr::holes()) \%>\% tail()
dt <- dplyr::bind_rows(holes_1d_better, holes_1d_geo)
dt \%>\%  bind_theoretical(best, tourr::holes()) \%>\% tail()
}
