% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{botanical_palettes}
\alias{botanical_palettes}
\alias{botanical_pal}
\alias{scale_color_botanical}
\title{A customised color palette based on australian botanicals}
\format{
An object of class \code{list} of length 5.
}
\usage{
botanical_palettes

botanical_pal(palette = "fern", reverse = FALSE)

scale_color_botanical(..., palette = "fern", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Color palette from the botanical_palette}

\item{reverse}{logical, if the color should be reversed}

\item{...}{Arguments passed into scale_color_gradientn}

\item{discrete}{logical, whether use a discrete or continuous color}
}
\description{
A customised color palette based on australian botanicals

Color interpolation for botanical palettes

A scale function in compatible with ggplot2
}
\keyword{datasets}
