% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{botanical_pal}
\alias{botanical_pal}
\title{Color interpolation}
\usage{
botanical_pal(palette = "fern", reverse = FALSE)
}
\arguments{
\item{palette}{Color palette from the botanical_palette}

\item{reverse}{logical, if the color should be reversed}
}
\description{
Color interpolation
}
\seealso{
Other botan: 
\code{\link{botanical_palettes}},
\code{\link{scale_color_botanical}()}
}
\concept{botan}
