% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{botanical_palettes}
\alias{botanical_palettes}
\title{A customised color palette based on australian botanicals}
\format{
An object of class \code{list} of length 5.
}
\usage{
botanical_palettes
}
\description{
Available colors in the palettes
}
\seealso{
Other botan: 
\code{\link{botanical_pal}()},
\code{\link{scale_color_botanical}()},
\code{\link{scale_fill_botanical}()}
}
\concept{botan}
\keyword{datasets}
