% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-smoothness.R
\name{calc_smoothness}
\alias{calc_smoothness}
\title{Function to calculate smoothness}
\usage{
calc_smoothness(
  idx,
  data = sine1000,
  size = 300,
  n = 6,
  d = 2,
  best = matrix(c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1), nrow = 6),
  start_parms = c(0.001, 0.5, 2, 2),
  other_gp_params = list(NULL)
)
}
\arguments{
\item{idx}{character, the name of projection pursuit index function, e.g.
"holes" (see the \pkg{tourr} package for index examples)}

\item{data}{matrix, the data to be projected}

\item{size}{numeric, the number of random basis to generate for calculating smoothness}

\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}

\item{best}{a matrix, the theoretical best projection matrix, used to calculate
projection distance with the simulated random bases.}

\item{start_parms}{Optionally specified starting values for parameters. 
If \code{NULL},
fit_model will select default starting values.}

\item{other_gp_params}{list, other parameters to be passed to \code{GpGp::fit_model}}
}
\description{
Function to calculate smoothness
}
