% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{botanical_palettes}
\alias{botanical_palettes}
\alias{botanical_pal}
\title{A customised colour palette based on Australian botanies}
\format{
An object of class \code{list} of length 5.
}
\usage{
botanical_palettes

botanical_pal(palette = "fern", reverse = FALSE)
}
\arguments{
\item{palette}{Colour palette from the botanical_palette}

\item{reverse}{logical, if the colour should be reversed}
}
\value{
a function for interpolating colour in the botanical palette
}
\description{
Available colours in the palettes
}
\keyword{datasets}
