% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute PCA for the projection bases}
\usage{
compute_pca(dt, group = NULL, random = TRUE, flip = TRUE, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{random}{Boolean, if the random data from the high dimensional sphere should be bounded}

\item{flip}{boolean, if the sign flipping need to be computed}

\item{...}{other arguments passed}
}
\description{
Compute PCA for the projection bases
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\% compute_pca(group = method)
}
