% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes_1d_geo}
\alias{holes_1d_geo}
\alias{holes_1d_better}
\alias{holes_2d_better}
\alias{holes_2d_better_max_tries}
\alias{boa_all}
\alias{boa5}
\alias{boa6}
\title{Simulated data}
\format{
The code for simulating each data object is as follows:
\preformatted{
set.seed(123456);
holes_1d_geo <-
  animate_dist(boa5, tour_path = guided_tour(holes(), d = 1,
                                             search_f =  search_geodesic),
               rescale = FALSE, verbose = TRUE)

set.seed(123456)
holes_1d_better <-
  animate_dist(boa5, tour_path = guided_tour(holes(), d = 1,
                                             search_f =  search_better),
               rescale = FALSE, verbose = TRUE)


set.seed(123456)
holes_2d_better <-
  animate_xy(boa6, tour_path = guided_tour(holes(), d = 2,
                                           search_f =  search_better),
            rescale = FALSE, verbose = TRUE)

set.seed(123456)
holes_2d_better_max_tries <-
  animate_xy(boa6, tour_path = guided_tour(holes(), d = 2,
                                           search_f = search_better,
                                           max.tries = 500),
             rescale = FALSE, verbose = TRUE)


}












}
\usage{
holes_1d_geo

holes_1d_better

holes_2d_better

holes_2d_better_max_tries

boa_all

boa5

boa6
}
\description{
The \code{boa} data is simulated using different gaussian mixtures with varied centers and weights,
whilst those prefixed \code{holes_*} are data objects generated by the tourr package (see \keyword{format} for the code of simulation).
}
\details{
Boa is given as the name of the dataset because the density plot of each variable in the data looks like boa constrictors swallowing multiple french sticks, rather than elephants, as per the little prince.

\code{holes} indicates the usage of holes index, \code{1d} or \code{2d} represent if the five variable \code{boa5} or the six variable \code{boa6} is used
The third chunk of the name indicates the seraching method used i.e. \code{geo} for geodesic search and \code{better} for \code{search_better}
Simulated data using different optimisation algorithms in guided tour.
}
\examples{

boa_all \%>\%
 pivot_longer(cols = x1: x10, names_to = "var", values_to = "value") \%>\%
 ggplot(aes(x = value)) +
 geom_density() +
 facet_wrap(vars(var))
}
\keyword{datasets}
