% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_anno}
\alias{draw_anno}
\title{A ggproto for annotating the symmetry of the starting points}
\usage{
draw_anno(dt, anno_color = "black", anno_lty = "dashed", anno_alpha = 0.1)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{anno_color}{the color of the annotation}

\item{anno_lty}{the linetype of the annotation}

\item{anno_alpha}{an alpha value for the transparency of the annotation}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_center}()},
\code{\link{draw_circle}()},
\code{\link{draw_path}()},
\code{\link{draw_points}()},
\code{\link{draw_theo}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
