% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_center}
\alias{draw_center}
\title{A ggproto for drawing the circle center}
\usage{
draw_center(dt, cent_size = 1, cent_alpha = 1, cent_color = "black")
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{cent_size}{the center point size}

\item{cent_alpha}{an alpha value for the transparency of the center point}

\item{cent_color}{the color of the center}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_circle}()},
\code{\link{draw_path}()},
\code{\link{draw_points}()},
\code{\link{draw_theo}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
