% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_circle}
\alias{draw_circle}
\title{A ggproto for drawing circle}
\usage{
draw_circle(dt, cir_alpha = 0.5, cir_fill = "grey92", cir_color = "white")
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{cir_alpha}{an alpha value for the transparency of the circle}

\item{cir_fill}{the color of the circle filling}

\item{cir_color}{the color of the circle brim}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_center}()},
\code{\link{draw_path}()},
\code{\link{draw_points}()},
\code{\link{draw_theo}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
