% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_path}
\alias{draw_path}
\title{A ggproto for drawing the path}
\usage{
draw_path(
  dt,
  path_size = 1.5,
  path_alpha = NULL,
  path_color = NULL,
  path_group = NULL,
  ...
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{path_size}{the size of the path}

\item{path_alpha}{an alpha value for the transparency of the path}

\item{path_color}{the color of the path}

\item{path_group}{a group variable for path connection}

\item{...}{other argument passed to \code{draw_path()}}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_center}()},
\code{\link{draw_circle}()},
\code{\link{draw_points}()},
\code{\link{draw_theo}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
