% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_points}
\alias{draw_points}
\title{A ggproto for drawing points}
\usage{
draw_points(dt, pnt_size = 1, pnt_alpha = 1, pnt_color = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{pnt_size}{the point size}

\item{pnt_alpha}{an alpha value for the transparency of the point}

\item{pnt_color}{the color of the points}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_center}()},
\code{\link{draw_circle}()},
\code{\link{draw_path}()},
\code{\link{draw_theo}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
