% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_theo}
\alias{draw_theo}
\title{A ggproto for drawing the theoretical basis, if applicable}
\usage{
draw_theo(dt, theo_label = "*", theo_size = 25)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{theo_label}{the symbol used for labelling the theoretical basis}

\item{theo_size}{the size of the label}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_center}()},
\code{\link{draw_circle}()},
\code{\link{draw_path}()},
\code{\link{draw_points}()},
\code{\link{estimate_circle}()}
}
\concept{draw functions}
