% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{estimate_circle}
\alias{estimate_circle}
\title{Estimate the radius of the background circle based on the randomly generated points}
\usage{
estimate_circle(dt)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}
}
\description{
The space of projected bases is a circle when reduced to 2D. A radius is estimated using
the largest distance from the bases in the data object to the center point.
}
\details{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{draw_anno}()},
\code{\link{draw_center}()},
\code{\link{draw_circle}()},
\code{\link{draw_path}()},
\code{\link{draw_points}()},
\code{\link{draw_theo}()}
}
\concept{draw functions}
