% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_proj_pca.R
\name{compute_pca}
\alias{compute_pca}
\alias{explore_proj_pca}
\title{explore projected basis in 2D}
\usage{
compute_pca(glb_obj)

explore_proj_pca(glb_obj, col = info, animate = FALSE)
}
\arguments{
\item{glb_obj}{a global object resulted from the \href{https://cran.r-project.org/web/packages/tourr}{tourr} package}

\item{animate}{a boolean value indicating whether to animate the projected basis across id}
}
\value{
a ggplot object plotting the 2D projection of the basis
}
\description{
explore projected basis in 2D
}
\examples{
\dontrun{holes_1d_geo \%>\% explore_proj_pca()}
\dontrun{bind_rows(holes_1d_geo, holes_1d_geo_polish) \%>\% explore_proj_pca(col = method)}
}
