% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{compute_pca}
\alias{compute_pca}
\alias{explore_space_pca}
\alias{explore_space_tour}
\title{Explore the parameter space}
\usage{
compute_pca(dt, group = NULL, random = TRUE)

explore_space_pca(dt, pca = TRUE, group = NULL, color = NULL, animate = FALSE)

explore_space_tour(dt, color = info, pal = botanical_palettes$banksia, ...)
}
\arguments{
\item{dt}{A data object to plot}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{random}{Boolean, if the random data from the high dimensional sphere should be bounded}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{animate}{Whether the plot should be animated}

\item{pal}{Color palette for \code{explore_space_tour()}}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo), color = method)
}
