% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{flip_sign}
\alias{flip_sign}
\alias{compute_pca}
\alias{explore_space_pca}
\title{Flip the sign of a group of bases}
\usage{
flip_sign(dt, group = NULL, ...)

compute_pca(dt, group = NULL, random = TRUE, flip = TRUE, ...)

explore_space_pca(
  dt,
  details = FALSE,
  pca = TRUE,
  group = NULL,
  color = NULL,
  ...,
  animate = FALSE
)
}
\arguments{
\item{dt}{a data object collected by the projection pursuit guided tour optimisation in \code{tourr}}

\item{group}{the variable to label different runs of the optimiser(s)}

\item{...}{other arguments passed to \code{add_*()} functions}

\item{random}{logical; if random bases from the basis space need to be added to the data}

\item{flip}{logical; if the sign flipping need to be performed}

\item{details}{logical; if components other than start, end and interpolation need to be shown}

\item{pca}{logical; if PCA coordinates need to be computed for the data}

\item{color}{the variable to be coloured by}

\item{animate}{logical; if the interpolation path needs to be animated}
}
\value{
a list containing 1) all the bases in a matrix format,
2) whether a flip of sign is performed, and 3) the original dataset supplied

a list containing 1) the PCA summary and 2) the data object after augmenting the PC coordinate

a ggplot object for diagnosing the optimisers in the PCA-projected basis space
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  flip_sign(group = method) \%>\%
  str(max = 1)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\% compute_pca(group = method)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
    index = tourr::holes(), raw_data = boa5
  ) \%>\%
  explore_space_pca(group = method, details = TRUE) +
  scale_color_discrete_botanical()
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
