% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{explore_space_tour}
\alias{explore_space_tour}
\title{Plot the grand tour animation of the projection bases space}
\usage{
explore_space_tour(dt, color = info, palette = botanical_palettes$banksia)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{palette}{The color palette to use}
}
\description{
Plot the grand tour animation of the projection bases space
}
\examples{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo), color = method)
}
\seealso{
Other plot: 
\code{\link{explore_space_pca}()},
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()}
}
\concept{plot}
