% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace_all.R
\name{explore_trace_all}
\alias{explore_trace_all}
\alias{explore_trace_interp}
\alias{explore_trace_search}
\title{explore trace plot}
\usage{
explore_trace_all(glb_obj, col = info, magnify = FALSE)

explore_trace_interp(glb_obj, point = FALSE)

explore_trace_search(glb_obj, col = alpha)
}
\arguments{
\item{glb_obj}{a global object resulted from the \href{https://cran.r-project.org/web/packages/tourr}{tourr} package}

\item{magnify}{a boolean value indicating whether to magnify the index value for the last \code{tries}}
}
\value{
a ggplot object plotting the trace
}
\description{
explore trace plot
}
\details{
* `explore_trace_all()`: plots all the index value from the global object as a dot plot.

* `explore_trace_interp()`: plots a line plot for the index value of the interpolated basis.
}
\examples{
\dontrun{
explore_trace_all(x2_object, 7)
explore_trace_interp(x2_object)
}
}
