% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R, R/myexplore_trace.R
\name{explore_trace_search}
\alias{explore_trace_search}
\title{Plot the count in each iteration}
\usage{
explore_trace_search(
  dt,
  iter = tries,
  color = tries,
  cutoff = 15,
  group = NULL
)

explore_trace_search(
  dt,
  iter = tries,
  color = tries,
  cutoff = 15,
  group = NULL
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Colored by a particular variable}

\item{cutoff}{The cutoff number of observations for switching between point geom to boxplot geom in \code{explore_trace_search()}}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object to plot}

\item{extend_lower}{a numeric for extending the y-axis to display text labels}

\item{...}{arguments passed into geom_label_repel() for displaying text labels}
}
\value{
a ggplot object for diagnosing how many points the optimiser(s) have searched
}
\description{
The purpose of this plot is to examine the variation in index values
obtained in each iteration. If there are just a few bases sampled these
will be displayed as a dotplot, otherwise with more than 6 sampled
bases a boxplot is used. Colour indicates iteration.
}
\examples{
# Summary plots for search points in two algorithms
library(patchwork)
library(dplyr)
library(ggplot2)
p1 <- holes_1d_better \%>\% explore_trace_search() +
  scale_color_continuous_botanical(palette = "fern") +
  xlim(c(1, 9))
p2 <- holes_2d_better_max_tries \%>\% explore_trace_search() +
  scale_color_continuous_botanical(palette = "daisy")
p1 / p2
# Summary plots for search points in two algorithms
proj_1D <- holes_1d_better \%>\% dplyr::mutate(proj = "1D")
proj_2D <- holes_2d_better_max_tries \%>\% dplyr::mutate(proj = "2D")
dplyr::bind_rows(proj_1D, proj_2D) \%>\% explore_trace_search(group = proj) +
 scale_color_botanical(palette = "daisy")
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{flip_sign}()},
\code{\link{prep_space_tour}()}

Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{flip_sign}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
