% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrn-package.R
\docType{package}
\name{ferrn-package}
\alias{ferrn}
\alias{ferrn-package}
\title{ferrn: Facilitate Exploration of touRR optimisatioN}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/ferrn}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/ferrn/issues}
}

}
\author{
\strong{Maintainer}: Sherry Zhang \email{Huize.Zhang@monash.edu}

}
\keyword{internal}
