% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{flip_sign}
\alias{flip_sign}
\title{Flip the sign of a group of bases}
\usage{
flip_sign(dt, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}
}
\description{
Flip the sign of a group of bases
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  flip_sign(group = method) \%>\%
  str(max = 1)
}
