% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_basis_matrix}
\alias{get_basis_matrix}
\title{Extract all the bases as a matrix}
\usage{
get_basis_matrix(dt)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}
}
\description{
Extract all the bases as a matrix
}
\examples{
head(get_basis_matrix(holes_1d_better), 5)
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_best}()},
\code{\link{get_center}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
