% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_best}
\alias{get_best}
\title{get the basis/ index value}
\usage{
get_best(dt)
}
\arguments{
\item{dt}{a data object from the \href{https://cran.r-project.org/web/packages/tourr}{tourr} package}
}
\value{
a list with the best index value along with its basis
}
\description{
get the basis/ index value
}
\examples{
\dontrun{get_best(holes_1d_better)$basis}
\dontrun{get_best(holes_1d_better)$index_val}
}
