% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_center}
\alias{get_center}
\title{Extract the center point of the random circle from the starting points}
\usage{
get_center(dt, pca = FALSE, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{...}{other argument passed to \code{compute_pca()}}
}
\description{
Extract the center point of the random circle from the starting points
}
\examples{
dplyr::bind_rows(holes_1d_better, holes_1d_geo) \%>\% get_center(pca = TRUE, group = method)
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
