% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_dir_search_transformed}
\alias{get_dir_search_transformed}
\title{Extract directional search points during the optimisation}
\usage{
get_dir_search_transformed(dt, ratio = 3, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{ratio}{a buffer value to allow directional search points being distinguishable from the anchor points}

\item{...}{other argument passed to \code{compute_pca()}}
}
\description{
Extract directional search points during the optimisation
}
\examples{
holes_1d_geo \%>\% compute_pca() \%>\% purrr::pluck("aug") \%>\% get_dir_search_transformed()
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt_finish}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
