% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_interp}
\alias{get_interp}
\title{Extract interpolated records}
\usage{
get_interp(dt, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}
}
\description{
Extract interpolated records
}
\examples{
holes_1d_better \%>\% get_interp()
get_interp(dplyr::bind_rows(holes_1d_better, holes_1d_geo), group = method)
}
\seealso{
Other get functions: 
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_search_count}()},
\code{\link{get_start}()}
}
\concept{get functions}
