% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_interrupt}
\alias{get_interrupt}
\title{Extract the end point of the interpolation and the target point in the iteration when an interruption happens}
\usage{
get_interrupt(dt, group = NULL, precision = 0.01, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}
}
\description{
The optimiser can find better basis on the interpolation path, an interruption is
implemented to stop further interpolation from the highest point to the target point.
This discrepancy is highlighted in the PCA plot. You should not use geodesic search on this function.
}
\examples{
holes_1d_better \%>\% get_interrupt()
holes_1d_geo \%>\% get_interrupt()
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_dir_search_transformed}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt_finish}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
