% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_interrupt_finish}
\alias{get_interrupt_finish}
\title{Extract the end point of the interpolation on the iteration where interruption happens}
\usage{
get_interrupt_finish(dt)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}
}
\description{
The optimiser can find better basis on the interpolation path, an interruption is
implemented to stop further interpolation from the highest point to the target point.
This discrepancy is highlighted in the PCA plot. You should not use geodesic search on this function.
}
\examples{
holes_1d_better \%>\% get_interrupt_finish()
holes_1d_geo \%>\% get_interrupt_finish()
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_center}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
