% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_search}
\alias{get_search}
\title{Extract search points during the optimisation}
\usage{
get_search(dt)
}
\arguments{
\item{dt}{a data object collected by the projection pursuit guided tour optimisation in the \code{tourr} package}
}
\value{
a tibble object containing the search bases
}
\description{
Extract search points during the optimisation
}
\examples{
holes_1d_better \%>\% get_search()
holes_1d_geo \%>\% get_search()
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_dir_search}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
