% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_search_count}
\alias{get_search_count}
\title{Extract the count in each iteration}
\usage{
get_search_count(dt, iter = NULL, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The variable used to be counted by}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}
}
\description{
Extract the count in each iteration
}
\examples{
get_search_count(holes_1d_better)
get_search_count(dplyr::bind_rows(holes_1d_better, holes_1d_geo), group = method)
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_center}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt_finish}()},
\code{\link{get_interrupt}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
