% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_space_param}
\alias{get_space_param}
\title{Estimate the radius of the background circle based on the randomly generated points}
\usage{
get_space_param(dt)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}
}
\description{
The space of projected bases is a circle when reduced to 2D. A radius is estimated using
the largest distance from the bases in the data object to the center point.
}
\details{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_center}()},
\code{\link{get_dir_search_transformed}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt_finish}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_start}()},
\code{\link{get_theo}()}
}
\concept{get functions}
