% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_start}
\alias{get_start}
\title{Extract the starting records}
\usage{
get_start(dt)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}
}
\description{
Extract the starting records
}
\examples{
holes_1d_better \%>\% get_start()
}
\seealso{
Other get functions: 
\code{\link{get_anchor}()},
\code{\link{get_basis_matrix}()},
\code{\link{get_best}()},
\code{\link{get_center}()},
\code{\link{get_interp_last}()},
\code{\link{get_interp}()},
\code{\link{get_interrupt}()},
\code{\link{get_search_count}()},
\code{\link{get_search}()},
\code{\link{get_space_param}()},
\code{\link{get_theo}()}
}
\concept{get functions}
