% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-smoothness.R, R/calc-squintability.R
\name{calc_smoothness}
\alias{calc_smoothness}
\alias{calc_squintability}
\alias{fit_ks}
\alias{fit_nls}
\title{Function to calculate smoothness and squintability}
\usage{
calc_smoothness(
  idx,
  data = sine1000,
  n_basis = 300,
  n = 6,
  d = 2,
  best = matrix(c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1), nrow = 6),
  start_parms = c(0.001, 0.5, 2, 2),
  other_gp_params = NULL
)

calc_squintability(
  idx,
  data = sine1000,
  return_early = FALSE,
  method = c("ks", "nls"),
  n_basis = 50,
  n = 6,
  d = 2,
  proj_dist_threshold = 1.5,
  step = 0.005,
  best = matrix(c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1), nrow = 6),
  bin_nobs_threshold = 5000,
  bin_width = 0.005,
  sampling_seed = 123
)

fit_ks(basis_df, idx, bin_nobs_threshold = 5000, bin_width = 0.005)

fit_nls(
  basis_df,
  idx,
  bin_nobs_threshold = 5000,
  bin_width = 0.005,
  nls_params = list(start = list(theta1 = 1, theta2 = 1, theta3 = 50, theta4 = 0))
)
}
\arguments{
\item{idx}{character, the name of projection pursuit index function, e.g.
"holes" (see the \pkg{tourr} package for index examples)}

\item{data}{matrix, the data to be projected}

\item{n_basis}{numeric, the number of random basis to generate for calculating smoothness}

\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}

\item{best}{a matrix, the theoretical best projection matrix, used to calculate
projection distance with the simulated random bases.}

\item{start_parms}{Optionally specified starting values for parameters. 
If \code{NULL},
fit_model will select default starting values.}

\item{other_gp_params}{list, other parameters to be passed to \code{GpGp::fit_model}}

\item{return_early}{logical, whether to return early of all the bases
before fitting the kernel or non-linear least square. This can be useful if
the index value evaluation is time-consuming and the user wants to save a copy
before fitting the kernel or non-linear least square.}

\item{method}{only for squintability, the method to calculate squintability,
either through kernel smoothing ("ks") or non-linear least square ("nls")}

\item{proj_dist_threshold}{only for squintability, the threshold for projection
distance for the random basis to be considered in sampling}

\item{step}{only for squintability, the step size for interpolation,
recommend between 0.01 and 0.1}

\item{bin_nobs_threshold}{numeric, only for squintability, the threshold
number of observations for
applying binning before fitting the kernel}

\item{bin_width}{only for squintability, the bin size for binning the data
before fitting the kernel, recommend to set as the same as step parameter}

\item{sampling_seed}{the seed used for sampling the random basis}

\item{basis_df}{a basis data frame, returned from \code{calc_squintability
(..., return_early = TRUE)}}

\item{nls_params}{additional parameter for fitting the nls model, see
\code{stats::nls()}}
}
\description{
Function to calculate smoothness and squintability
}
