% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_info.R
\name{relevel_geo}
\alias{relevel_geo}
\alias{relevel_better}
\alias{clean_method}
\title{Clean factor levels in the data object}
\usage{
relevel_geo(
  dt,
  order = c("new_basis", "direction_search", "best_direction_search",
    "best_line_search", "interpolation")
)

relevel_better(dt, order = c("random_search", "new_basis", "interpolation"))

clean_method(dt)
}
\arguments{
\item{dt}{a data object}

\item{order}{A vector that allows you to customise the order of info}
}
\description{
Relevel the \code{info} column in the data object to be the step in the optimisation algorithms.
}
\details{
For \code{search_geodesic},
\itemize{
\item{new_basis < direction_search < best_direction_search < best_line_search}}

For \code{search_better} and \code{search_better_random},

\itemize{
\item{random_search < new_basis < interpolation}}
}
\examples{
holes_1d_geo \%>\% relevel_geo()
holes_1d_better \%>\% relevel_better()
}
