% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_botanical}
\alias{scale_color_botanical}
\title{Scale color function in compatible with ggplot2 syntax}
\usage{
scale_color_botanical(..., palette = "fern", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{...}{Arguments passed into scale_color_gradientn}

\item{palette}{Color palette from the botanical_palette}

\item{discrete}{logical, whether use a discrete or continuous color}

\item{reverse}{logical, if the color should be reversed}
}
\description{
Scale color function in compatible with ggplot2 syntax
}
\seealso{
Other botan: 
\code{\link{botanical_palettes}},
\code{\link{botanical_pal}()}
}
\concept{botan}
