% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_fill_discrete_botanical}
\alias{scale_fill_discrete_botanical}
\title{discrete scale fill function}
\usage{
scale_fill_discrete_botanical(palette = "fern", reverse = FALSE)
}
\arguments{
\item{palette}{Color palette from the botanical_palette}

\item{reverse}{logical, if the color should be reversed}
}
\description{
discrete scale fill function
}
\seealso{
Other botan: 
\code{\link{botanical_palettes}},
\code{\link{botanical_pal}()},
\code{\link{scale_color_continuous_botanical}()},
\code{\link{scale_color_discrete_botanical}()},
\code{\link{scale_fill_continuous_botanical}()}
}
\concept{botan}
