% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_anno}
\alias{add_anno}
\title{A ggproto for annotating the symmetry of the starting points}
\usage{
add_anno(dt, anno_color = "black", anno_lty = "dashed", anno_alpha = 0.1, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{anno_color}{character; the colour of the annotation line}

\item{anno_lty}{character; the linetype of the annotation line}

\item{anno_alpha}{numeric; the alpha of the annotation line}

\item{...}{other aesthetics inherent from \code{explore_space_pca()}}
}
\value{
a wrapper for annotating the symmetry of start points in \code{explore_space_pca()}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_dir_search}()},
\code{\link{add_end}()},
\code{\link{add_interp}()},
\code{\link{add_interp_last}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
