% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_search}
\alias{add_search}
\title{A ggproto for drawing search points}
\usage{
add_search(dt, search_size = 0.5, search_alpha = 0.5, search_color = NULL, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{search_size}{numeric; the size of the search points}

\item{search_alpha}{numeric; the alpha of the anchor points}

\item{search_color}{the variable to be coloured by}

\item{...}{other aesthetics inherent from \code{explore_space_pca()}}
}
\value{
a wrapper for drawing search points in \code{explore_space_pca()}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_dir_search}()},
\code{\link{add_end}()},
\code{\link{add_interp}()},
\code{\link{add_interp_last}()},
\code{\link{add_interrupt}()},
\code{\link{add_space}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
