% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_space}
\alias{add_space}
\title{A ggproto for drawing circle}
\usage{
add_space(dt, cir_alpha = 0.5, cir_fill = "grey92", cir_color = "white")
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{cir_alpha}{an alpha value for the transparency of the circle}

\item{cir_fill}{the color of the circle filling}

\item{cir_color}{the color of the circle brim}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_finish}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_start}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
