% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{add_start}
\alias{add_start}
\title{A ggproto for drawing start points}
\usage{
add_start(dt, start_size = 5, start_alpha = 1, start_color = NULL, ...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{start_size}{the point size}

\item{start_alpha}{an alpha value for the transparency of the point}

\item{start_color}{the color of the points}
}
\description{
This is  a wrapper function used by \code{explore_space_pca()} and
should be be called directly by the user
}
\seealso{
Other draw functions: 
\code{\link{add_anchor}()},
\code{\link{add_anno}()},
\code{\link{add_center}()},
\code{\link{add_finish}()},
\code{\link{add_interp}()},
\code{\link{add_interrupt}()},
\code{\link{add_search}()},
\code{\link{add_space}()},
\code{\link{add_theo}()}
}
\concept{draw functions}
