% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_theoretical}
\alias{bind_theoretical}
\title{Bind the theoretical best record}
\usage{
bind_theoretical(dt, matrix, index, raw_data)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{matrix}{The theoretical basis to bind}

\item{index}{The index function used to calculate index value}

\item{raw_data}{The original data, used to compute the index value for the theoretical best basis}
}
\description{
The theoretical best basis is usually known for a simulated problem.
Augment this information into the data object allows for evaluating the performance of optimisation against the theory.
}
\examples{
best <- matrix(c(0, 1, 0, 0, 0), nrow = 5)
tail(holes_1d_better \%>\% bind_theoretical(best, tourr::holes(), raw_data = boa5), 1)
}
\seealso{
Other bind: 
\code{\link{bind_random_matrix}()},
\code{\link{bind_random}()}
}
\concept{bind}
