% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes_1d_geo}
\alias{holes_1d_geo}
\alias{holes_1d_better}
\alias{holes_2d_better}
\alias{holes_2d_better_max_tries}
\alias{data}
\alias{data_mult}
\title{Dataset}
\format{
All the dataset uses the same data format as described in the paper.

\code{holes_1d_geo}: 1D projection using \code{search_geodesic} in guided tour in a simulated 5-variable dataset

\code{holes_1d_better}: 1D projection using \code{search_better} in a simulated 5-variable dataset

\code{holes_2d_better}: 2D projection using \code{search_better}
in a simulated 6-variable dataset

\code{holes_2d_better}: 2D projection using \code{search_better}
in a simulated 6-variable dataset while allow the maximum number
of tries to be sufficiently large




}
\usage{
holes_1d_geo

holes_1d_better

holes_2d_better

holes_2d_better_max_tries

data

data_mult
}
\description{
Simulated data using different optimisation algorithms in guided tour
}
\keyword{datasets}
