% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes_1d_geo}
\alias{holes_1d_geo}
\alias{holes_1d_better}
\alias{holes_1d_jellyfish}
\alias{holes_2d_better}
\alias{holes_2d_better_max_tries}
\title{Data objects collected during the projection pursuit optimisation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 416 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 79 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2500 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 98 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1499 rows and 8 columns.
}
\usage{
holes_1d_geo

holes_1d_better

holes_1d_jellyfish

holes_2d_better

holes_2d_better_max_tries
}
\description{
Simulated data to demonstrate the usage of four diagnostic plots in the package,
users can create their own guided tour data objects and diagnose with the visualisation designed in this package.
}
\details{
The prefix \code{holes_*} indicates the use of holes index in the guided tour.
The suffix \code{*_better/geo/jellyfish} indicates the optimiser used:
 \code{search_better}, \code{search_geodesic}, \code{search_jellyfish}.
}
\examples{
holes_1d_better \%>\%
explore_trace_interp(interp_size = 2) +
  scale_color_continuous_botanical(palette = "fern")
}
\keyword{datasets}
