% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{compute_pca}
\alias{compute_pca}
\alias{explore_space_pca}
\alias{explore_space_tour}
\title{Plotting the data object in the space reduced by PCA}
\usage{
compute_pca(dt, random = TRUE)

explore_space_pca(dt, pca = TRUE, color = info, animate = FALSE, ...)

explore_space_tour(dt, color = info, pal = pal, ...)
}
\arguments{
\item{dt}{a data object to plot}

\item{pca}{a data object after performing PCA}

\item{...}{other arguments passed into \code{explore_space_pca}}

\item{col}{the color of the point}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
best <- matrix(c(0, 1, 0, 0, 0), nrow = 5)
with_theo <- bind_theoretical(holes_1d_better, best, tourr::holes())
pca <- compute_pca(with_theo)$aug
explore_space_pca(pca)
}
