% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R, R/myexplore_space.R
\name{flip_sign}
\alias{flip_sign}
\alias{compute_pca}
\alias{explore_space_pca}
\title{Flip the sign of a group of bases}
\usage{
flip_sign(dt, group = NULL)

compute_pca(dt, group = NULL, random = TRUE)

explore_space_pca(dt, pca = TRUE, group = NULL, color = NULL, animate = FALSE)

explore_space_pca(dt, pca = TRUE, group = NULL, color = NULL, animate = FALSE)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{random}{logical; if random bases from the basis space need to be added to the data}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{animate}{Boolean, if the plot should be animated}

\item{flip}{logical; if the sign flipping need to be performed}

\item{details}{logical; if components other than start, end and interpolation need to be shown}

\item{...}{other arguments passed to \code{add_*()} functions}
}
\value{
a list containing 1) all the bases in a matrix format,
2) whether a flip of sign is performed, and 3) the original dataset supplied

a list containing 1) the PCA summary and 2) the data object after augmenting the PC coordinate

a ggplot object for diagnosing the optimisers in the PCA-projected basis space
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`

The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  flip_sign(group = method) \%>\%
  str(max = 1)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\% compute_pca(group = method)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
    index = tourr::holes(), raw_data = boa5
  ) \%>\%
  explore_space_pca(group = method, details = TRUE) +
  scale_color_discrete_botanical()
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
 bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
                  index = tourr::holes(), raw_data = boa5) \%>\%
 explore_space_pca(group = method)  +
 scale_color_botanical(palette = "cherry")
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}

Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
