% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{explore_space_pca}
\alias{explore_space_pca}
\title{Plot the PCA projection of the projection bases space}
\usage{
explore_space_pca(dt, pca = TRUE, group = NULL, color = NULL, animate = FALSE)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{animate}{Whether the plot should be animated}

\item{pal}{Color palette for \code{explore_space_tour()}}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\seealso{
Other plot: 
\code{\link{explore_space_tour}()},
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()}
}
\concept{plot}
