% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{explore_space_pca}
\alias{explore_space_pca}
\title{Plot the PCA projection of the projection bases space}
\usage{
explore_space_pca(
  dt,
  pca = TRUE,
  group = NULL,
  color = NULL,
  cir_alpha = 0.5,
  cir_fill = "grey92",
  cir_color = "white",
  cent_size = 1,
  cent_alpha = 1,
  cent_color = "black",
  start_size = 5,
  start_alpha = 1,
  anchor_size = 3,
  anchor_alpha = 1,
  search_size = 0.5,
  search_alpha = 0.5,
  interp_size = 1.5,
  interrupt_size = 0.5,
  anno_color = "black",
  anno_lty = "dashed",
  anno_alpha = 0.1,
  theo_size = 25,
  theo_label = "*",
  animate = FALSE
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{cir_alpha}{an argument passed to \code{add_circle()}}

\item{cir_fill}{an argument passed to \code{add_circle()}}

\item{cir_color}{an argument passed to \code{add_circle()}}

\item{cent_size}{an argument passed to \code{add_center()}}

\item{cent_alpha}{an argument passed to \code{add_center()}}

\item{cent_color}{an argument passed to \code{add_center()}}

\item{start_size}{an argument passed to \code{add_start()}}

\item{start_alpha}{an argument passed to \code{add_start()}}

\item{anchor_size}{an argument passed to \code{add_anchor()}}

\item{anchor_alpha}{an argument passed to \code{add_anchor()}}

\item{search_size}{an argument passed to \code{add_search()}}

\item{search_alpha}{an argument passed to \code{add_search()}}

\item{interp_size}{an argument passed to \code{add_interp()}}

\item{interrupt_size}{an argument passed to \code{add_interrupt()}}

\item{anno_color}{an argument passed to \code{add_anno()}}

\item{anno_lty}{an argument passed to \code{add_anno()}}

\item{anno_alpha}{an argument passed to \code{add_anno()}}

\item{theo_size}{an argument passed to \code{add_theo()}}

\item{theo_label}{an argument passed to \code{add_theo()}}

\item{animate}{Boolean, if the plot should be animated}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
    index = tourr::holes(), raw_data = boa5
  ) \%>\%
  explore_space_pca(group = method) +
  scale_color_botanical(palette = "cherry")
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
