% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{flip_sign}
\alias{flip_sign}
\alias{compute_pca}
\alias{explore_space_pca}
\title{Flip the sign of a group of bases}
\usage{
flip_sign(dt, group = NULL, ...)

compute_pca(dt, group = NULL, random = TRUE, flip = TRUE, ...)

explore_space_pca(
  dt,
  details = FALSE,
  pca = TRUE,
  group = NULL,
  color = NULL,
  ...,
  animate = FALSE
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{...}{different argument passed to \code{add_*()}}

\item{random}{Boolean, if the random data from the high dimensional sphere should be bounded}

\item{flip}{boolean, if the sign flipping need to be computed}

\item{details}{if all the components of the optimisers should be shown. With \code{details = FALSE} only the start, end and interpolation path will be shown}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{color}{A variable from the object that the diagnostic plot should be coloured by}

\item{animate}{Boolean, if the plot should be animated}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  flip_sign(group = method) \%>\%
  str(max = 1)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\% compute_pca(group = method)
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
    index = tourr::holes(), raw_data = boa5
  ) \%>\%
  explore_space_pca(group = method, details = TRUE) +
  scale_color_discrete_botanical()
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
