% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{explore_space_pca}
\alias{explore_space_pca}
\title{Plot the PCA projection of the projection bases space}
\usage{
explore_space_pca(
  dt,
  pca = TRUE,
  group = NULL,
  color = NULL,
  ...,
  animate = FALSE
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{pca}{Boolean, if \code{compute_pca()} should be performed on the data}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{...}{different argument passed to \code{add_*()}}

\item{animate}{Boolean, if the plot should be animated}
}
\description{
The set of functions returns a primary ggplot object
that plots the data object in a space reduced by PCA.
\code{compute_pca()} computes the PCA and \code{explore_space_pca()} does the plotting.`
}
\examples{
dplyr::bind_rows(holes_1d_geo, holes_1d_better) \%>\%
  bind_theoretical(matrix(c(0, 1, 0, 0, 0), nrow = 5),
    index = tourr::holes(), raw_data = boa5
  ) \%>\%
  explore_space_pca(group = method) +
  scale_color_botanical(palette = "cherry")
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
