% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{prep_space_tour}
\alias{prep_space_tour}
\alias{explore_space_tour}
\title{Plot the grand tour animation of the projection bases space}
\usage{
prep_space_tour(
  dt,
  group = NULL,
  color = NULL,
  rand_size = 1,
  point_size = 1.5,
  end_size = 5,
  theo_size = 3,
  theo_shape = 17,
  palette = botanical_palettes$fern,
  ...
)

explore_space_tour(...)
}
\arguments{
\item{dt}{a data object collected by the projection pursuit guided tour optimisation in \code{tourr}}

\item{group}{the variable to label different runs of the optimiser(s)}

\item{color}{the variable to be coloured by}

\item{rand_size}{numeric; the size of random points}

\item{point_size}{numeric; the size of points searched by the optimiser(s)}

\item{end_size}{numeric; the size of end points}

\item{theo_size}{numeric; the size of theoretical point(s)}

\item{theo_shape}{numeric; the shape symbol in the basic plot}

\item{palette}{the colour palette to be used}

\item{...}{other argument passed to \code{tourr::animate_xy()}}
}
\value{
a list containing various components needed for producing the tour plot

a tour animation of the search path in the original high-dimensional basis space
}
\description{
Plot the grand tour animation of the projection bases space
}
\examples{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo),
  group = method, palette = botanical_palettes$fern[c(1, 6)]
)
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()}
}
\concept{plot}
