% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R, R/myexplore_space.R
\name{prep_space_tour}
\alias{prep_space_tour}
\alias{explore_space_tour}
\title{Plot the grand tour animation of the projection bases space}
\usage{
prep_space_tour(
  dt,
  group = NULL,
  flip = FALSE,
  color = NULL,
  rand_size = 1,
  point_size = 1.5,
  end_size = 5,
  theo_size = 3,
  theo_shape = 17,
  theo_color = "black",
  palette = botanical_palettes$fern,
  ...
)

explore_space_tour(
  dt,
  group = NULL,
  theoretical = FALSE,
  color = method,
  palette = botanical_palettes$cherry,
  ...
)

explore_space_tour(
  dt,
  group = NULL,
  theoretical = FALSE,
  color = method,
  palette = botanical_palettes$cherry,
  ...
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{flip}{logical; if the sign flipping need to be performed}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{rand_size}{numeric; the size of random points}

\item{point_size}{numeric; the size of points searched by the optimiser(s)}

\item{end_size}{numeric; the size of end points}

\item{theo_size}{numeric; the size of theoretical point(s)}

\item{theo_shape}{numeric; the shape symbol in the basic plot}

\item{theo_color}{character; the color of theoretical point(s)}

\item{palette}{The color palette to use}

\item{...}{Additional argument passed to \code{tourr::animate_xy()}}

\item{theoretical}{Boolean, if the theoretical bases have been inlcuded in the data object}
}
\value{
a list containing various components needed for producing the tour plot

a tour animation of the search path in the original high-dimensional basis space
}
\description{
Plot the grand tour animation of the projection bases space

Plot the grand tour animation of the projection bases space
}
\examples{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo),
  group = method, palette = botanical_palettes$fern[c(1, 6)]
)
\dontrun{explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo), group = method)}
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()}

Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()}
}
\concept{plot}
