% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{prep_space_tour}
\alias{prep_space_tour}
\alias{explore_space_tour}
\title{Plot the grand tour animation of the projection bases space}
\usage{
prep_space_tour(
  dt,
  group = NULL,
  theoretical = FALSE,
  color = sym("method"),
  rand_size = 9,
  point_size = 1.5,
  theo_size = 10,
  palette = botanical_palettes$cherry,
  ...
)

explore_space_tour(...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{theoretical}{Boolean, if the theoretical bases have been inlcuded in the data object}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{rand_size}{random point size}

\item{point_size}{other point size}

\item{theo_size}{theoretical point size}

\item{palette}{The color palette to use}

\item{...}{Additional argument passed to \code{tourr::animate_xy()}}
}
\description{
Plot the grand tour animation of the projection bases space
}
\examples{
\dontrun{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo), group = method)
}
}
\seealso{
Other plot: 
\code{\link{explore_space_pca}()},
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()}
}
\concept{plot}
