% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_space.R
\name{prep_space_tour}
\alias{prep_space_tour}
\alias{explore_space_tour}
\title{Plot the grand tour animation of the projection bases space}
\usage{
prep_space_tour(
  dt,
  group = NULL,
  color = NULL,
  rand_size = 1,
  point_size = 1.5,
  end_size = 5,
  theo_size = 3,
  theo_shape = 17,
  palette = botanical_palettes$fern,
  ...
)

explore_space_tour(...)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{color}{A variable from the object that the diagnostic plot should be colored by}

\item{rand_size}{random point size}

\item{point_size}{other point size}

\item{end_size}{the size of the ending point}

\item{theo_size}{theoretical point size}

\item{theo_shape}{the numerical value of basic plot shape}

\item{palette}{The color palette to use}

\item{...}{Additional argument passed to \code{tourr::animate_xy()}}
}
\description{
Plot the grand tour animation of the projection bases space
}
\examples{
\dontrun{
explore_space_tour(dplyr::bind_rows(holes_1d_better, holes_1d_geo),
  group = method, palette = botanical_palettes$fern[c(1, 6)]
)
}
}
\seealso{
Other plot: 
\code{\link{explore_trace_interp}()},
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()}
}
\concept{plot}
