% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R
\name{explore_trace_interp}
\alias{explore_trace_interp}
\title{Plot the trace the search progression}
\usage{
explore_trace_interp(
  dt,
  iter = NULL,
  color = NULL,
  group = NULL,
  cutoff = 50,
  target_size = 3,
  interp_size = 1,
  accuracy_x = 5,
  accuracy_y = 0.01
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Coloured by a particular variable}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}

\item{cutoff}{If there are less than cut-off number of points on the interpolation path, all the points will be marked}

\item{target_size}{The size of the target points in the interpolation}

\item{interp_size}{The size of the interpolation points}

\item{accuracy_x}{if two x neighbour values are closer than accuracy_x, only one of them will be displayed. Used for better axis label}

\item{accuracy_y}{the precision of y-axis label}
}
\description{
Trace the index value of search/ interpolation points in guided tour optimisation
}
\examples{
# Compare the trace of interpolated points in two algorithms
holes_1d_better \%>\%
  explore_trace_interp() +
  scale_color_continuous_botanical(palette = "fern")
}
\seealso{
Other plot: 
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
