% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R, R/myexplore_trace.R
\name{explore_trace_interp}
\alias{explore_trace_interp}
\title{Plot the trace the search progression}
\usage{
explore_trace_interp(dt, iter = id, color = tries, group = NULL)

explore_trace_interp(dt, iter = id, color = tries, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Colored by a particular varaible}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object to plot}

\item{cutoff}{numeric; if the number of interpolating points is smaller than \code{cutoff}, all the interpolation points will be plotted as dots}

\item{target_size}{numeric; the size of target points in the interpolation}

\item{interp_size}{numeric; the size of interpolation points}

\item{accuracy_x}{numeric; If the difference of two neighbour x-labels is smaller than \code{accuracy_x}, only one of them will be displayed. Used for better axis label}

\item{accuracy_y}{numeric; the precision of y-axis label}
}
\value{
a ggplot object for diagnosing how the index value progresses during the interpolation
}
\description{
Trace the index value of search/ interpolation points in guided tour optimisation

This plot shows the trace of the index value, through
the search and interpolation between best projections
for a guided tour optimisation.
}
\examples{
# Compare the trace of interpolated points in two algorithms
holes_1d_better \%>\%
  explore_trace_interp() +
  scale_color_continuous_botanical(palette = "fern")
 # Compare the trace of interpolated points in two algorithms
 dplyr::bind_rows(holes_1d_better, holes_1d_geo) \%>\%
   explore_trace_interp(group = method) +
   scale_color_botanical(palette = "fern")
}
\seealso{
Other plot: 
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()},
\code{\link{prep_space_tour}()}

Other plot: 
\code{\link{explore_trace_search}()},
\code{\link{flip_sign}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
