% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R
\name{explore_trace_interp}
\alias{explore_trace_interp}
\title{Plot the trace the search progression}
\usage{
explore_trace_interp(dt, iter = id, color = tries, fill = tries)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Colored by a particular varaible}
}
\description{
Trace the index value of search/ interpolation points in guided tour optimisation
}
\examples{
# Compare the trace of interpolated points in two algorithms
holes_1d_better \%>\%
 explore_trace_interp() +
 scale_color_botanical(palette = "fern", discrete = FALSE) +
 scale_fill_botanical(palette = "fern", discrete = FALSE)
}
\seealso{
Other plot: 
\code{\link{explore_space_pca}()},
\code{\link{explore_space_tour}()},
\code{\link{explore_trace_search}()}
}
\concept{plot}
