% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace_all.R
\name{explore_trace_interp}
\alias{explore_trace_interp}
\alias{explore_trace_search}
\title{trace the progression of the search}
\usage{
explore_trace_interp(dt, iter = id, col = tries, group = NULL)

explore_trace_search(dt, iter = tries, col = tries, cutoff = 15, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{col}{Colored by a particular varaible}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object to plot}

\item{cutoff}{The cutoff number of observations for switching between point geom to boxplot geom in \code{explore_trace_search()}}
}
\description{
trace the index value of search/ interpolation points in guided tour optimisation
}
\examples{
# Summary plots for search points in two algorithms
1D <- holes_1d_better \%>\% mutate(proj = "1D"),
2D <- holes_2d_better_max_tries \%>\% mutate(proj = "2D")
bind_rows(1D, 2D) \%>\% explore_trace_search(group = proj)

# Compare the trace of interpolated points in two algorithms
interp <- bind_rows(holes_1d_better, holes_1d_geo) \%>\% get_interp(group = method)
interp \%>\% explore_trace_interp(group = method)
}
