% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R
\name{explore_trace_search}
\alias{explore_trace_search}
\title{Plot the count in each iteration}
\usage{
explore_trace_search(dt, iter = tries, color = tries, cutoff = 15)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Colored by a particular varaible}

\item{cutoff}{The cutoff number of observations for switching between point geom to boxplot geom in \code{explore_trace_search()}}
}
\description{
Plot the count in each iteration
}
\examples{
# Summary plots for search points in two algorithms
holes_1d_better \%>\% explore_trace_search() +
 scale_color_botanical(palette = "fern")
holes_2d_better_max_tries \%>\% explore_trace_search() +
 scale_color_botanical(palette = "daisy")
}
\seealso{
Other plot: 
\code{\link{explore_space_pca}()},
\code{\link{explore_space_tour}()},
\code{\link{explore_trace_interp}()}
}
\concept{plot}
