% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_trace.R
\name{explore_trace_search}
\alias{explore_trace_search}
\title{Plot the count in each iteration}
\usage{
explore_trace_search(
  dt,
  iter = NULL,
  color = NULL,
  cutoff = 15,
  extend_lower = 0.95,
  ...
)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{iter}{The iterator on the x-axis}

\item{color}{Colored by a particular varaible}

\item{cutoff}{The cutoff number of observations for switching between point geom to boxplot geom in \code{explore_trace_search()}}

\item{extend_lower}{a percentage to extend the lower ylim for better display of the count}

\item{...}{arguments passed into geom_label_repel() for better display of the count in each iteration}
}
\description{
Plot the count in each iteration
}
\examples{
# Summary plots for search points in two algorithms
library(patchwork)
library(dplyr)
library(ggplot2)
p1 <- holes_1d_better \%>\% explore_trace_search() +
  scale_color_continuous_botanical(palette = "fern") +
  xlim(c(1, 9))
p2 <- holes_2d_better_max_tries \%>\% explore_trace_search() +
  scale_color_continuous_botanical(palette = "daisy")
p1 / p2
}
\seealso{
Other plot: 
\code{\link{explore_space_pca}()},
\code{\link{explore_trace_interp}()},
\code{\link{prep_space_tour}()}
}
\concept{plot}
