% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrn-package.R
\docType{package}
\name{ferrn-package}
\alias{ferrn}
\alias{ferrn-package}
\title{ferrn: Facilitate Exploration of touRR optimisatioN}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

`ferrn` serves to produce diagnostic plots for projection pursuit guided tour optimisers . 
    It also provides functionality for extracting various components in the data objects collected in `tourr`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/ferrn/}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/ferrn/issues}
}

}
\author{
\strong{Maintainer}: H. Sherry Zhang \email{huize.zhang@monash.edu} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID})

Authors:
\itemize{
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID})
  \item Nicolas Langrené \email{nicolas.langrene@csiro.au} (\href{https://orcid.org/0000-0001-7601-4618}{ORCID})
  \item Patricia Menéndez \email{patricia.menendez@monash.edu} (\href{https://orcid.org/0000-0003-0701-6315}{ORCID})
}

}
\keyword{internal}
