% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrn-package.R
\docType{package}
\name{ferrn-package}
\alias{ferrn}
\alias{ferrn-package}
\title{ferrn: Facilitate Exploration of touRR optimisatioN}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A package for extracting key components in the data object collected by the guided tour optimisation, and producing diagnostic plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/ferrn/}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/ferrn/issues}
}

}
\author{
\strong{Maintainer}: Sherry Zhang \email{huize.zhang@monash.edu}

}
\keyword{internal}
