% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_label}
\alias{format_label}
\title{Better label formatting to avoid overlapping}
\usage{
format_label(labels, accuracy)
}
\arguments{
\item{labels}{a numerical vector of labels}

\item{accuracy}{the accuracy of the label}
}
\value{
a vector of adjusted labels
}
\description{
Better label formatting to avoid overlapping
}
\examples{
format_label(c(0.87, 0.87, 0.9, 0.93, 0.95), 0.01)
format_label(c(0.87, 0.87, 0.9, 0.93, 0.95, 0.96, 0.96), 0.01)
}
