% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_best}
\alias{get_best}
\alias{get_start}
\alias{get_interp}
\alias{get_search_count}
\title{get the basis/ index value}
\usage{
get_best(dt, group = NULL)

get_start(dt)

get_interp(dt, group = NULL)

get_search_count(dt, iter = tries, group = NULL)
}
\arguments{
\item{dt}{A data object from the running the optimisation algorithm in guided tour}

\item{group}{The grouping variable, useful when there are multiple algorithms in the data object}
}
\description{
This set of functions allows you to grab the best basis found by the optimisation in guided tour
as well as the bases on the interpolation path.
}
\examples{
holes_1d_better \%>\% get_start()
holes_1d_better \%>\% get_interp()
bind_rows(holes_1d_better, holes_1d_geo) \%>\% get_best(group = method)
}
