% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-projection.R
\name{plot_projection}
\alias{plot_projection}
\alias{compute_projection}
\title{Plot the projection from the optimisation data collected from projection pursuit}
\usage{
plot_projection(dt, data, cols = NULL)

compute_projection(dt, data, cols = NULL)
}
\arguments{
\item{dt}{a data object collected by the projection pursuit guided tour optimisation in \code{tourr}}

\item{data}{the original data}

\item{cols}{additional columns to include in the plot}
}
\value{
a ggplot object
}
\description{
Plot the projection from the optimisation data collected from projection pursuit
}
\examples{
holes_1d_jellyfish |> get_best() |> plot_projection(data = boa5)
}
