% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_method}
\alias{clean_method}
\title{Clean method names}
\usage{
clean_method(dt)
}
\arguments{
\item{dt}{a data object}
}
\value{
a tibble with method cleaned
}
\description{
Clean method names
}
\examples{
head(clean_method(holes_1d_better), 5)
}
